added createOrder api

+++ src/main/java/com/mishri/order_service_ecommerce/dto/OrderItemRequestDTO.java
@@ -9,6 +9,5 @@ import lombok.*;
 @Builder
 public class OrderItemRequestDTO {
     private Long productId;
-
-    private Long quantity;
+    private Integer quantity;
 }
